move Udongo::Search::ResultObject to Udongo::Search::ResultObjects::Base
